% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informativeSCI-package.R
\docType{package}
\name{informativeSCI-package}
\alias{informativeSCI}
\alias{informativeSCI-package}
\title{informativeSCI: Informative Simultaneous Confidencde Intervals}
\description{
Calculation of informative simultaneous confidencde intervals for graphical tests and exploration of the information weight.
}
\details{
The main function of the package is the `informSCI`-function for
calculating informative lower simultaneous confidence bounds for a given
graphical test procedure and given information weights. The
`explore_q`-function can help to find the right information weights
for the \link{informSCI}-algorithm. The `inExactSCI`- and
`notInExactSCI`-functions can help to determine how accurate a
(numerical) approximation of the true informative lower SCI-bounds is.
}
\references{
F. Bretz, W. Maurer, W. Brannath, M. Posch: A graphical approach
to sequentially rejective multiple test procedures. Statistics in Medicine
28.4 (2009), pp. 586-604.

K. Strassburger, F. Bretz: Compatible simultaneous lower confidence bounds
for the Holm procedure and other Bonferroni based closed tests. Statistics
in Medicine 27.4 (2008), pp. 4914-4927.

W. Brannath, L. Kluge, M. Scharpenberg: Informative Simultaneous Confidence 
Intervals for Graphical Test Procedures. arXiv preprint arXiv:2402.13719 (2024).

S. Schmidt, W. Brannath: Informative Simultaneous Confidence Intervals
in Hierarchical Testing. Methods of Information in Medicine 53.4 (2014),
pp. 278–283.
}
\author{
\strong{Maintainer}: Kluge Liane \email{liane@uni-bremen.de}

Other contributors:
\itemize{
  \item Brannath Werner [contributor]
  \item Scharpenberg Martin [contributor]
}

}
